Avoid writing empty abbrev tables
Fixes bug#29923

'insert-abbrev-table-description' with a non-nil READABLE inserts Lisp
forms suitable for evaluation to restore the defined abbrevs.  We
don't have to insert a form for tables that do not have any abbrevs.

To implement this, we need to filter out system abbrevs before
checking if a table is empty, because system abbrevs were previously
skipped in the 'abbrev--write' call, at which point we would already
have started inserting the beginning of a table definition form.

* lisp/abbrev.el (insert-abbrev-table-description):
Skip inserting empty tables when READABLE is non-nil.
Clarify behavior in documentation string.
(abbrev--write): Remove system abbrev check.

* doc/lispref/abbrevs.texi (Abbrev Tables): Document behavior
with empty tables.

* etc/NEWS: Mention the change in behavior of
'insert-abbrev-table-description'.