(WIP) add a lisp_fn_result macro that uses an explicit Result return type
This is a work-in-progress because I did the lazy thing and just
copied the lisp_fn macro to make the lisp_fn_result version. Either we
should share the bulk of the implementation between them, or we should
make all Lisp functions written in Rust use Results, or we should
ignore this branch.