simplest factoring of the common elements of the two macros
Honestly I don't really like this. Looking at `lisp_fn` and
`lisp_fn_result` you can't really tell _why_ the code is written the
way it is. You can't see where or how `ret` is declared, or how it
gets its value. Looking at `lisp_fn_internal`, you only see an opaque
blob of `Tokens`, with no real clue as to what they are doing. This
makes it harder to understand the overall effect of either macro. I'm
pretty certain that this will get worse as these get more complicated,
and as more variants are added, but I suppose it's not too bad at the
moment. And it's better than completly duplicating the entire macro.