Ensure woman2-roff-buffer restores functions on error (Bug#30908)
* lisp/woman.el (woman2-roff-buffer): Put the fallback paragraph
processing inside the protected part of the unwind-protect form,
rather than the cleanup forms.  Attempting to format paragraphs again
after an error has been signaled is unlikely to be helpful.  The
fallback processing should be triggered only in case the loop
terminated normally, but did not reach the end of the buffer for some
reason.