Add tools for git bisect
Use like this. Pay attention to the escaping!!

$ pwd
/path/to/top/of/remacs/tree
$ git bisect start
$ git bisect good @{"2 weeks ago"}  # last time you know it worked
$ git bisect bad HEAD               # top of current tree
$ git bisect run rust_src/admin/did-it-work.sh -Q --eval "\"(some-lisp)\""
Each iteration the script will build remacs and run it with the
arguments provided. Exit emacs if possible or Control-C it from the
terminal. You will then be asked "Did it work?". Press Y or N and
enter. Git will find the first broken commit if possible.

The broken commit hash will be output by Git if it found it. Run git bisect reset when you are done and the tree will be back to where you started. git checkout ABC123 # the bad hash will set the tree to the first broken version so you can investigate.