Amend c-before-change-check-unbalanced-strings to handle a quote at EOB.
* lisp/progmodes/cc-mode.el (c-before-change-check-unbalanced-strings): Before
going back a sexp from the \n marked with a string-fence syntax-table text
property, check that it is indeed a \n.  If not, simply delete that text
property from the isolated quote, not the quote and the \n which otherwise
delimit an unterminated string.  Should we encounter a double quote marked
with a string-fence syntax-table text property at the end of the last line,
remove that property.