Improve --with-wide-int mpz_t→fixnum conversion
These tuneups and minor simplifications should affect only
platforms with EMACS_INT wider than ‘long’.
* src/alloc.c (make_number): If the number fits in long but
not in fixnum, do not attempt to convert to fixnum again.
Tighten the compile-time check for whether the second attempt
is worth trying, from sizeof (long) < sizeof (EMACS_INT) to
LONG_WIDTH < FIXNUM_BITS.  Do not bother computing the sign of
the value to tighten the bounds for whether to try the second
attempt, as it’s not worth the effort.  Do not call mpz_size,
which is unnecessary since the number of bits is already known
and the loop can iterate over a shift count instead.  Avoid
unnecessary casts.  Use + instead of | where either will do,
as + is typically better for optimization.

Improve mpz_t to fixnum when --with-wide-int
* src/alloc.c (make_number): Avoid undefined behavior
when shifting an EMACS_UINT by more than EMACS_UINT_WIDTH bits.
Check for integer overflow when shifting.