Add support in emacsclient for EMACS_SOCKET_NAME
If the '--socket-name' argument is unspecified, the environment
variable 'EMACS_SOCKET_NAME' is now consulted with the same
semantics.  This mirrors the behavior of the '--server-file' argument,
and allows for easier configuration of emacsclient when the socket is
in a location other than 'TMPDIR' or '/tmp'.

* emacsclient.c (set_socket): Add support for the
EMACS_SOCKET_NAME environment variable.  (Bug#33095)

* misc.texi (emacsclient Options):
* emacsclient.1: Document the EMACS_SOCKET_NAME environment
variable.

* etc/NEWS: Announce the new feature.

Copyright-paperwork-exempt: yes