; * etc/NEWS: assoc-delete-all fix
This is a mess.  This function was added in master in afba4ccb8b8,
undocumented.  It was then tweaked and documented in 9824885fabe.
Much later in cc233365a92 the original untweaked and undocumented
version was backported to emacs-26.  This was independently
redocumented in emacs-26 in cc8f334d2da, which conflicted with the
interface changes that had been made on master.