Remove many .rustified_enum calls
Of the (non-windows-gated) calls that are still required, the
reason almost all are required is simply that the type name is
mentioned directly in rust_src. They could be changed to
`[module_name]::Type` if desired. However, in the case of
`output_method`, there's a `match` at
https://github.com/Wilfred/remacs/blob/master/rust_src/src/frames.rs#L161
that needs exhaustivity checking, which the `ModuleConsts` enum style
doesn't provide, since it just make the enum into an int instead of a
rust enum. Is there another reason why
`.default_enum_style(bindgen::EnumVariation::ModuleConsts)` is
desirable that I'm not aware of, as opposed to having the default
be `bindgen::EnumVariation::Rust`?

Also conservatively removed the windows `.rustified_enums` that don't
actually occur anywhere in `rust_src/`.